✨ Generate inputs, filters,sorts and resolvers from entity models #9
