+++ src/db-reverse/generation/model-generation.ts
@@ -11,6 +11,7 @@ import IGenerationOptions, {
 import { Entity } from "../models/entity";
 import { Relation } from "../models/relation";
 import { singular } from "pluralize";
+import { ModelGenerationOptions } from '../../commands/generate/crud';
 
 const prettierOptions: Prettier.Options = {
   parser: "typescript",
@@ -46,7 +47,8 @@ export default function modelGenerationPhase(
 
 export function modelGenerationCodeFirst(
   generationOptions: IGenerationOptions,
-  databaseModel: Entity[]
+  databaseModel: Entity[],
+  flags: ModelGenerationOptions
 ) {
   createHandlebarsHelpers(generationOptions);
   const resultPath = generationOptions.resultsPath;
@@ -60,13 +62,13 @@ export function modelGenerationCodeFirst(
       fs.mkdirSync(entitiesPath);
     }
   }
-  generateGraphQLFiles(databaseModel, generationOptions, entitiesPath);
+  generateGraphQLFiles(databaseModel, generationOptions, entitiesPath, flags);
 }
 
 function generateFiles(
   databaseModel: Entity[],
   generationOptions: IGenerationOptions,
-  entitiesPath: string
+  entitiesPath: string  
 ) {
   const entityTemplatePath = path.resolve(
     __dirname,
@@ -192,7 +194,8 @@ function generateFiles(
 function generateGraphQLFiles(
   databaseModel: Entity[],
   generationOptions: IGenerationOptions,
-  entitiesPath: string
+  entitiesPath: string,
+  flags: ModelGenerationOptions
 ) { 
   const filtersTemplatePath = path.resolve(
     __dirname,
@@ -261,39 +264,50 @@ function generateGraphQLFiles(
 
     fs.mkdirSync(filesPathModels, { recursive: true });
     fs.mkdirSync(filesPathResolvers, { recursive: true });
-      
-    generateFilters(
-      databaseModel,
-      generationOptions,
-      baseFileName,
-      filesPathModels,
-      filtersCompliedTemplate,
-      element
-    );
-    generateSort(
-      databaseModel,
-      generationOptions,
-      baseFileName,
-      filesPathModels,
-      sortsCompliedTemplate,
-      element
-    );
-    generateInput(
-      databaseModel,
-      generationOptions,
-      baseFileName,
-      filesPathModels,
-      inputsCompliedTemplate,
-      element
-    );
-    generateResolver(
-      databaseModel,
-      generationOptions,
-      baseFileName,
-      filesPathResolvers,
-      resolverCompliedTemplate,
-      element
-    );
+     
+    if(flags.filter){
+      generateFilters(
+        databaseModel,
+        generationOptions,
+        baseFileName,
+        filesPathModels,
+        filtersCompliedTemplate,
+        element
+      );
+    }
+    
+    if(flags.sort){
+      generateSort(
+        databaseModel,
+        generationOptions,
+        baseFileName,
+        filesPathModels,
+        sortsCompliedTemplate,
+        element
+      );
+    }
+
+    if(flags.input){
+      generateInput(
+        databaseModel,
+        generationOptions,
+        baseFileName,
+        filesPathModels,
+        inputsCompliedTemplate,
+        element
+      );
+    }
+
+    if(flags.resolver){
+      generateResolver(
+        databaseModel,
+        generationOptions,
+        baseFileName,
+        filesPathResolvers,
+        resolverCompliedTemplate,
+        element
+      );
+    }   
   });
 }
